# contracts/openapi/chat2db-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat2DB API",
    "description": "Transport contract for the Chat2DB Rust Community runtime",
    "contact": {
      "name": "OtterMind"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/datasources": {
      "get": {
        "tags": [
          "datasources"
        ],
        "operationId": "list_datasources",
        "responses": {
          "200": {
            "description": "Secret-free datasource list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceList"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected datasource failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Datasource storage is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasources"
        ],
        "operationId": "create_datasource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Datasource created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datasource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid datasource request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Datasource conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected datasource failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Datasource storage is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/{datasource_id}": {
      "get": {
        "tags": [
          "datasources"
        ],
        "operationId": "get_datasource",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "description": "Opaque datasource id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Secret-free datasource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datasource"
                }
              }
            }
          },
          "404": {
            "description": "Datasource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected datasource failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Datasource storage is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "datasources"
        ],
        "operationId": "update_datasource",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "description": "Opaque datasource id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatasourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Datasource updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datasource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid datasource request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Datasource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Datasource revision conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected datasource failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Datasource storage is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasources"
        ],
        "operationId": "delete_datasource",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "description": "Opaque datasource id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expectedRevision",
            "in": "query",
            "description": "Expected monotonic revision",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Datasource deleted"
          },
          "400": {
            "description": "Invalid expected revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Datasource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Datasource revision conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected datasource failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Datasource storage is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operations/{operation_id}": {
      "get": {
        "tags": [
          "operations"
        ],
        "operationId": "operation_snapshot",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "description": "Opaque operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current operation snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "Operation does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/{operation_id}/cancel": {
      "post": {
        "tags": [
          "operations"
        ],
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "description": "Opaque operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Idempotent cancellation disposition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/{operation_id}/events": {
      "get": {
        "tags": [
          "operations"
        ],
        "operationId": "operation_events",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "description": "Opaque operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "description": "Last consumed numeric event sequence",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replay followed by live operation events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Last-Event-ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Operation does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Requested event fell outside the replay window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected subscription failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Stable error envelope used at every external transport boundary.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiErrorDetails",
                "description": "Optional, bounded diagnostic context."
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "Safe user-facing message."
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether the caller may retry this operation under the same policy."
          }
        }
      },
...
      "CancelOperationResponse": {
        "type": "object",
        "description": "Cancellation response shared by Web and desktop transports.",
        "required": [
          "operationId",
          "disposition"
        ],
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/CancelDisposition",
            "description": "Idempotent cancellation disposition."
          },
          "operationId": {
            "type": "string",
            "description": "Opaque operation id supplied by the caller."
          }
        }
      },
...
      "CreateDatasourceRequest": {
        "type": "object",
        "description": "Request to create one datasource and optionally install its connection secret.",
        "required": [
          "name",
          "driverId"
        ],
        "properties": {
          "connection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DatasourceConnection",
                "description": "Complete connection descriptor, omitted when credentials are installed later."
              }
            ]
          },
          "driverId": {
            "type": "string",
            "description": "Compatibility-engine driver id."
          },
          "name": {
            "type": "string",
            "description": "User-visible datasource name."
          }
        }
      },
...
      "DatasourceList": {
        "type": "object",
        "description": "Stable datasource collection returned by list APIs.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Datasource"
            },
            "description": "Datasources in stable creation order."
          }
        }
      },
...
    }
  },
...
}